v3.19.9 - UI workflow audit: Low-severity fixes
Resolves all 96 Low-severity findings from the UI workflow audit: 65 fixed in
code, 31 documented/scoped. Builds clean on Win32 and Win64 (D_D130, Release).
Priorities throughout: integrity, reliability, performance.

Reliability/lifetime: option-page frames validate SetUserData casts and nil-guard
Load/Save (Compile Backup, Compiler Enhancements, File Cleaner); AfterCompile
style-check wrapped so it cannot escape the IDE notifier; native progress label
nil-guards FProgressBar; debugger step-into-skip validates the dcc32 handle/exports
and uses GetModuleHandle (no leaked handle); Dependency Viewer close-during-scan
veto; Keybindings SetActive uses Supports not as; Path Sorter singleton constructed
safely + InputQuery; Unused Unit Detector closes its form on unload; External Mod
Monitor syncs watches on Active toggle, watches group members, and bounds-checks
the notify buffer; clipboard/icon-load guarded; Application.MainForm nil-guarded.

Correctness: Code Quality bracket-depth array-index exemption + memory-leak feature
disabled; Component Selector absolute best-fit; Dependency Viewer consistent cycle
counts + PATH parsing; Empty-Handler declaration line; Unreachable-Code truncation
flag; TODO numeric priority sort; CSV RFC-4180 escaping across five exporters;
File Cleaner Unicode LowerCase; Old Palette GetClass nil-guard.

Diagnosability/standards: silent hook failures and the reload-hook exception now
log; IDE menu handler re-resolves the Project menu each tick; Options URL label
reports failure; stray debug Write; removed; inline-var loop counter per GITLAK.

Version bumped 3.18.8 -> 3.19.9 (version.inc, version.h, 7 .dproj). CHANGES.md
updated; desktop audit doc reclassified to 221 fixed / 40 documented / 0 open /
21 refuted.

+++ Code/DDevExtensions/Source/ExcelExport/FrmExcelExport.pas
@@ -125,7 +125,7 @@ begin
         end;
         if Assigned(ProgressBar) and (Items.Count > 0) and (i mod 25 = 0) then
         begin
-          ProgressBar.Position := i * 100 div Items.Count;
+          ProgressBar.Position := Int64(i) * 100 div Items.Count;
           Application.ProcessMessages;
         end;
       end;
